pyproject.toml
[project]
name = "dis22-sose2026-movie-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.68.0",
    "openai>=2.29.0",
    "pandas>=2.3.3",
    "streamlit>=1.55.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "marimo>=0.21.1",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.7",
    "ty>=0.0.24",
]
